Made compiled component work
There must have been an old version of GenerateComponent.cs. on the remote, which was inconsistent with some of the renaming (mostly changing capitalisation) of the helper functions. Now GenerateComponent.cs is consistent with the renamed functions. Beware that GenerateComponent.cs should really just be a call to Generator() since compiled and scriptable algorithms should behave exactly the same.